Correcion

+++ ProyectoProgramacion2Unidad2/Form1.cs
@@ -50,10 +50,10 @@ namespace ProyectoProgramacion2Unidad2
             {
                 //if (dvgAlumnos.Rows[e.RowIndex].IsNewRow) return;
 
-                int resultado;
+                long resultado;
                 string valor = e.FormattedValue.ToString();
                 
-                if (!int.TryParse(valor, out resultado) || resultado < 0)
+                if (!long.TryParse(valor, out resultado) || resultado < 0)
                 {
                     e.Cancel = true;
                     MessageBox.Show("Solo se permite números enteros.", "Valor inválido", MessageBoxButtons.OK, MessageBoxIcon.Warning);
